feat: re-implemented spiral_data function in go

+++ core/layers.cs
@@ -7,7 +7,7 @@ namespace IDS_NN.core;
 	n_inputs: size of one record of the input data
 	n_neurons: number of neurons
 **/
-public class LayerDense(int n_inputs, int n_neurons)
+public class LayerDense(int n_inputs, ushort n_neurons)
 {
 	/**
 		multiply the _weights by the 0.10 to create a gaussian distribution bounded by zero
@@ -18,7 +18,7 @@ public class LayerDense(int n_inputs, int n_neurons)
 	public NDArray Output => _output ?? 0;
 	
 	
-	public void Forward(NDArray  inputs)
+	public void Forward(NDArray inputs)
 	{
 		_output = np.dot(inputs, _weights) + _biases;
 	}
